refactor(api): move remove-authentication-method usecase
in identity-access-management
